// CodeGenerator/models/station.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 서울 따릉이 시뮬레이션 플랫폼 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Station
 */
export interface Station {
    /**
     * 대여소 고유 ID
     * @type {string}
     * @memberof Station
     */
    'id': string;
    /**
     * 대여소 이름
     * @type {string}
     * @memberof Station
     */
    'name': string;
    /**
     * 대여소 주소
     * @type {string}
     * @memberof Station
     */
    'address': string;
    /**
     * 위도
     * @type {string}
     * @memberof Station
     */
    'latitude': string;
    /**
     * 경도
     * @type {string}
     * @memberof Station
     */
    'longitude': string;
    /**
     * 현재 대여 가능한 자전거 수
     * @type {number}
     * @memberof Station
     */
    'bike_count': number;
    /**
     * 대여소 최대 수용 대수
     * @type {number}
     * @memberof Station
     */
    'capacity': number;
    /**
     * 대여소 운영 상태 (active: 운영 중, inactive: 운영 중지)
     * @type {string}
     * @memberof Station
     */
    'status': StationStatusEnum;
}

export const StationStatusEnum = {
    Active: 'active',
    Inactive: 'inactive'
} as const;

export type StationStatusEnum = typeof StationStatusEnum[keyof typeof StationStatusEnum];