// CodeGenerator/models/bike.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 서울 따릉이 시뮬레이션 플랫폼 API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * 
 * @export
 * @interface Bike
 */
export interface Bike {
    /**
     * 
     * @type {string}
     * @memberof Bike
     */
    'id': string;
    /**
     * 
     * @type {string}
     * @memberof Bike
     */
    'status': BikeStatusEnum;
    /**
     * 
     * @type {string}
     * @memberof Bike
     */
    'current_station_id': string | null;
    /**
     * 
     * @type {string}
     * @memberof Bike
     */
    'created_at': string;
    /**
     * 
     * @type {string}
     * @memberof Bike
     */
    'updated_at': string;
    /**
     * 
     * @type {number}
     * @memberof Bike
     */
    'use_count': number;
}

export const BikeStatusEnum = {
    Available: 'available',
    Rented: 'rented',
    Maintenence: 'maintenence'
} as const;

export type BikeStatusEnum = typeof BikeStatusEnum[keyof typeof BikeStatusEnum];